Migrate NotificationPrefs screen to controller
