+++ presentation/src/main/res/layout/notification_prefs_presenter.xml
@@ -24,23 +24,9 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        style="@style/Toolbar"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <com.moez.QKSMS.common.widget.QkTextView
-            android:id="@+id/toolbarTitle"
-            style="@style/ToolbarText"
-            tools:text="Notifications" />
-
-    </androidx.appcompat.widget.Toolbar>
-
     <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/toolbar">
+        android:layout_height="match_parent">
 
         <LinearLayout
             android:id="@+id/preferences"
